refactor: 가격 정책을 enum class로 추가
