+++ app/src/main/java/woowacourse/movie/domain/PurchasePolicy.kt
@@ -1,0 +1,7 @@
+package woowacourse.movie.domain
+
+import java.time.LocalTime
+
+enum class PurchasePolicy(val time: LocalTime) {
+    NormalPolicy(LocalTime.of(0, 15)),
+}
