--- app/src/main/java/woowacourse/movie/domain/Ticket.kt
@@ -1,15 +1,0 @@
-package woowacourse.movie.domain
-
-import java.io.Serializable
-import java.time.LocalDateTime
-
-data class Ticket(
-    val title: String,
-    val date: LocalDateTime,
-    val personnel: Int,
-    val theaterName: String,
-) : Serializable {
-    companion object {
-        const val CANCEL_DEADLINE = 15
-    }
-}
