+++ app/src/main/java/woowacourse/movie/view/reservation/result/ReservationCompletePresenter.kt
@@ -1,8 +1,8 @@
 package woowacourse.movie.view.reservation.result
 
-import woowacourse.movie.domain.Ticket
 import woowacourse.movie.domain.movieseat.Position
 import woowacourse.movie.domain.movieseat.Seats
+import woowacourse.movie.view.reservation.Ticket
 
 class ReservationCompletePresenter(
     val view: ReservationCompleteContract.View,
