+++ app/src/main/java/woowacourse/movie/view/reservation/seat/ReservationSeatPresenter.kt
@@ -1,10 +1,10 @@
 package woowacourse.movie.view.reservation.seat
 
 import android.os.Bundle
-import woowacourse.movie.domain.Ticket
 import woowacourse.movie.domain.movieseat.Position
 import woowacourse.movie.domain.movieseat.Seat
 import woowacourse.movie.domain.movieseat.Seats
+import woowacourse.movie.view.reservation.Ticket
 
 class ReservationSeatPresenter(
     val view: ReservationSeatContract.View,
