+++ app/src/main/res/values/strings.xml
@@ -9,7 +9,7 @@
     <string name="default_personnel">0</string>
     <string name="minus">-</string>
     <string name="movie_title">영화 제목</string>
-    <string name="movie_cancel_deadline">영화 상영 시작 %1$d분 전까지\n 취소가 가능합니다.</string>
+    <string name="movie_cancel_deadline">영화 상영 시작 %1$s분 전까지\n 취소가 가능합니다.</string>
     <string name="default_movie_date_time">2025.1.1. 00:00</string>
     <string name="message_not_allowed_time">다른 날짜를 선택해 주세요</string>
     <string name="reserve_confirm">예매 확인</string>
