+++ app/src/test/java/woowacourse/movie/domain/ReservationSeatPresenterTest.kt
@@ -10,6 +10,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import woowacourse.movie.domain.movieseat.Position
+import woowacourse.movie.view.reservation.Ticket
 import woowacourse.movie.view.reservation.seat.ReservationSeatContract
 import woowacourse.movie.view.reservation.seat.ReservationSeatPresenter
 import java.time.LocalDateTime
